+++ jest.setup.js
@@ -1,1 +1,9 @@
-import '@testing-library/jest-dom'
+import '@testing-library/jest-dom';
+
+class ResizeObserverMock {
+  observe() {}
+  unobserve() {}
+  disconnect() {}
+}
+
+global.ResizeObserver = ResizeObserverMock;